refactor schema generation
