+++ ogcapi-draft/ogcapi-collections-queryables/src/main/java/de/ii/ldproxy/ogcapi/collections/queryables/app/QueryablesFormatExtension.java
@@ -8,12 +8,12 @@
 package de.ii.ldproxy.ogcapi.collections.queryables.app;
 
 import de.ii.ldproxy.ogcapi.common.domain.GenericFormatExtension;
+import de.ii.ldproxy.ogcapi.domain.ApiRequestContext;
 import de.ii.ldproxy.ogcapi.domain.Link;
 import de.ii.ldproxy.ogcapi.domain.OgcApi;
-import de.ii.ldproxy.ogcapi.domain.ApiRequestContext;
+import de.ii.ldproxy.ogcapi.features.geojson.domain.JsonSchemaObject;
 
 import java.util.List;
-import java.util.Map;
 
 public interface QueryablesFormatExtension extends GenericFormatExtension {
 
@@ -21,6 +21,6 @@ public interface QueryablesFormatExtension extends GenericFormatExtension {
         return "^\\/?collections\\/[^\\/]+\\/queryables/?$";
     }
 
-    Object getEntity(Map<String,Object> schemaQueryables, List<Link> links, String collectionId, OgcApi api, ApiRequestContext requestContext);
+    Object getEntity(JsonSchemaObject schemaQueryables, List<Link> links, String collectionId, OgcApi api, ApiRequestContext requestContext);
 
 }
