+++ ogcapi-stable/ogcapi-features-core/src/main/java/de/ii/ldproxy/ogcapi/features/core/app/EndpointFeatures.java
@@ -35,10 +35,11 @@ import de.ii.ldproxy.ogcapi.features.core.domain.FeaturesCoreQueriesHandler;
 import de.ii.ldproxy.ogcapi.features.core.domain.FeaturesQuery;
 import de.ii.ldproxy.ogcapi.features.core.domain.ImmutableQueryInputFeature;
 import de.ii.ldproxy.ogcapi.features.core.domain.ImmutableQueryInputFeatures;
+import de.ii.ldproxy.ogcapi.features.core.domain.SchemaGeneratorFeatureOpenApi;
 import de.ii.xtraplatform.auth.domain.User;
 import de.ii.xtraplatform.features.domain.FeatureQuery;
+import de.ii.xtraplatform.features.domain.FeatureSchema;
 import io.dropwizard.auth.Auth;
-import io.swagger.v3.oas.models.media.StringSchema;
 import org.apache.felix.ipojo.annotations.Component;
 import org.apache.felix.ipojo.annotations.Instantiate;
 import org.apache.felix.ipojo.annotations.Provides;
@@ -67,6 +68,9 @@ public class EndpointFeatures extends EndpointSubCollection {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(EndpointFeatures.class);
     private static final List<String> TAGS = ImmutableList.of("Access data");
+
+    @Requires
+    SchemaGeneratorFeatureOpenApi schemaGeneratorFeature;
 
     private final FeaturesCoreProviders providers;
     private final FeaturesQuery ogcApiFeaturesQuery;
@@ -195,6 +199,12 @@ public class EndpointFeatures extends EndpointSubCollection {
             // TODO
         }
 
+        List<FeatureSchema> featureProperties = providers.getFeatureProvider(apiData, featureType)
+                                                         .getData()
+                                                         .getTypes()
+                                                         .get(collectionId)
+                                                         .getProperties();
+
         if (definitionPath.equals("/collections/{collectionId}/items"))
             return ImmutableList.<OgcApiQueryParameter>builder()
                     .addAll(generalList)
@@ -204,7 +214,7 @@ public class EndpointFeatures extends EndpointSubCollection {
                                     .collectionId(collectionId)
                                     .name(field)
                                     .description("Filter the collection by property '" + field + "'")
-                                    .schema(new StringSchema())
+                                    .schema(schemaGeneratorFeature.getSchemaOpenApi(apiData, collectionId, field))
                                     .build())
                         .collect(Collectors.toList()))
                     .build();
