+++ ogcapi-stable/ogcapi-features-core/src/main/java/de/ii/ldproxy/ogcapi/features/core/domain/SchemaGeneratorFeature.java
@@ -1,0 +1,7 @@
+package de.ii.ldproxy.ogcapi.features.core.domain;
+
+public abstract class SchemaGeneratorFeature {
+
+    public enum SCHEMA_TYPE {QUERYABLES, RETURNABLES, RETURNABLES_FLAT, MUTABLES}
+
+}
