+++ ogcapi-stable/ogcapi-features-core/src/main/java/de/ii/ldproxy/ogcapi/features/core/domain/SchemaGeneratorFeatureCollectionOpenApi.java
@@ -1,21 +1,23 @@
-package de.ii.ldproxy.ogcapi.features.geojson.domain;
+package de.ii.ldproxy.ogcapi.features.core.domain;
 
 import com.google.common.collect.ImmutableList;
 import de.ii.ldproxy.ogcapi.domain.OgcApiDataV2;
-import de.ii.ldproxy.ogcapi.features.geojson.domain.SchemaGeneratorFeature;
-import io.swagger.v3.oas.models.media.*;
+import io.swagger.v3.oas.models.media.ArraySchema;
+import io.swagger.v3.oas.models.media.ObjectSchema;
+import io.swagger.v3.oas.models.media.Schema;
+import io.swagger.v3.oas.models.media.StringSchema;
 import org.apache.felix.ipojo.annotations.Component;
 import org.apache.felix.ipojo.annotations.Instantiate;
 import org.apache.felix.ipojo.annotations.Provides;
 import org.apache.felix.ipojo.annotations.Requires;
 
 @Component
-@Provides(specifications = {SchemaGeneratorFeatureCollection.class})
+@Provides(specifications = {SchemaGeneratorFeatureCollectionOpenApi.class})
 @Instantiate
-public class SchemaGeneratorFeatureCollection {
+public class SchemaGeneratorFeatureCollectionOpenApi {
 
     @Requires
-    SchemaGeneratorFeature schemaGeneratorFeature;
+    SchemaGeneratorFeatureOpenApi schemaGeneratorFeature;
 
     final static Schema GENERIC = new ObjectSchema()
             .required(ImmutableList.of("type","features"))
@@ -32,26 +34,26 @@ public class SchemaGeneratorFeatureCollection {
         return GENERIC;
     }
 
-    public String getSchemaReferenceOpenApi(String collectionId) { return "#/components/schemas/featureCollectionGeoJson_"+collectionId; }
+    public String getSchemaReferenceOpenApi(String collectionId, SchemaGeneratorFeature.SCHEMA_TYPE type) { return "#/components/schemas/featureCollectionGeoJson_"+collectionId; }
 
-    public Schema getSchemaOpenApi(OgcApiDataV2 apiData, String collectionId) {
+    public Schema getSchemaOpenApi(OgcApiDataV2 apiData, String collectionId, SchemaGeneratorFeature.SCHEMA_TYPE type) {
         return new ObjectSchema()
                 .required(ImmutableList.of("type","features"))
                 .addProperties("type", new StringSchema()._enum(ImmutableList.of("FeatureCollection")))
-                .addProperties("features", new ArraySchema().items(new Schema().$ref(schemaGeneratorFeature.getSchemaReferenceOpenApi(collectionId))))
+                .addProperties("features", new ArraySchema().items(new Schema().$ref(schemaGeneratorFeature.getSchemaReferenceOpenApi(collectionId, type))))
                 .addProperties("links", new ArraySchema().items(new Schema().$ref("https://raw.githubusercontent.com/opengeospatial/ogcapi-features/master/core/openapi/ogcapi-features-1.yaml#/components/schemas/link")))
                 .addProperties("timeStamp", new Schema().$ref("https://raw.githubusercontent.com/opengeospatial/ogcapi-features/master/core/openapi/ogcapi-features-1.yaml#/components/schemas/timeStamp"))
                 .addProperties("numberMatched", new Schema().$ref("https://raw.githubusercontent.com/opengeospatial/ogcapi-features/master/core/openapi/ogcapi-features-1.yaml#/components/schemas/numberMatched"))
                 .addProperties("numberReturned", new Schema().$ref("https://raw.githubusercontent.com/opengeospatial/ogcapi-features/master/core/openapi/ogcapi-features-1.yaml#/components/schemas/numberReturned"));
     }
 
-    public String getSchemaReferenceByName(String name) { return "#/components/schemas/featureCollectionGeoJson_"+name; }
+    public String getSchemaReferenceByName(String name, SchemaGeneratorFeature.SCHEMA_TYPE type) { return "#/components/schemas/featureCollectionGeoJson_"+name; }
 
-    public Schema getSchemaOpenApiForName(String name) {
+    public Schema getSchemaOpenApiForName(String name, SchemaGeneratorFeature.SCHEMA_TYPE type) {
         return new ObjectSchema()
                 .required(ImmutableList.of("type","features"))
                 .addProperties("type", new StringSchema()._enum(ImmutableList.of("FeatureCollection")))
-                .addProperties("features", new ArraySchema().items(new Schema().$ref(schemaGeneratorFeature.getSchemaReferenceOpenApi(name)))
+                .addProperties("features", new ArraySchema().items(new Schema().$ref(schemaGeneratorFeature.getSchemaReferenceOpenApi(name, type)))
                 .addProperties("links", new ArraySchema().items(new Schema().$ref("https://raw.githubusercontent.com/opengeospatial/ogcapi-features/master/core/openapi/ogcapi-features-1.yaml#/components/schemas/link")))
                 .addProperties("timeStamp", new Schema().$ref("https://raw.githubusercontent.com/opengeospatial/ogcapi-features/master/core/openapi/ogcapi-features-1.yaml#/components/schemas/timeStamp")))
                 .addProperties("numberMatched", new Schema().$ref("https://raw.githubusercontent.com/opengeospatial/ogcapi-features/master/core/openapi/ogcapi-features-1.yaml#/components/schemas/numberMatched"))
