+++ ogcapi-stable/ogcapi-features-geojson/src/main/java/de/ii/ldproxy/ogcapi/features/geojson/domain/JsonSchemaBoolean.java
@@ -1,0 +1,20 @@
+/**
+ * Copyright 2020 interactive instruments GmbH
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package de.ii.ldproxy.ogcapi.features.geojson.domain;
+
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import org.immutables.value.Value;
+
+@Value.Immutable
+@Value.Style(jdkOnly = true, deepImmutablesDetection = true)
+@JsonDeserialize(as = ImmutableJsonSchemaBoolean.class)
+public abstract class JsonSchemaBoolean extends JsonSchema {
+
+    public final String getType() { return "boolea"; }
+
+}
