+++ ogcapi-stable/ogcapi-features-geojson/src/main/java/de/ii/ldproxy/ogcapi/features/geojson/domain/JsonSchemaString.java
@@ -1,0 +1,28 @@
+/**
+ * Copyright 2020 interactive instruments GmbH
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package de.ii.ldproxy.ogcapi.features.geojson.domain;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import org.immutables.value.Value;
+
+import java.util.List;
+import java.util.Optional;
+
+@Value.Immutable
+@Value.Style(jdkOnly = true, deepImmutablesDetection = true)
+@JsonDeserialize(as = ImmutableJsonSchemaString.class)
+public abstract class JsonSchemaString extends JsonSchema {
+
+    public final String getType() { return "string"; }
+
+    public abstract Optional<String> getFormat();
+    public abstract Optional<String> getPattern();
+    @JsonProperty("enum")
+    public abstract List<String> getEnums();
+}
